+++ __init__.py
@@ -1,4 +1,4 @@
 from .converter import MicronConverter, default_url_resolver, UrlResolver, slugify_micron
 
 __all__ = ["MicronConverter", "default_url_resolver", "UrlResolver", "slugify_micron"]
-__version__ = "1.0.8"
+__version__ = "1.1.0"